feat: Add CommonDatePicker component for date selection
A new CommonDatePicker component has been added to facilitate date selection in forms. This component integrates the react-datepicker library and provides a user-friendly date picker interface. It has been implemented in the AppointmentModal component to allow users to select a date for their appointment. This enhancement improves the user experience and ensures accurate date input.