feat: Update AppointmentButton component to use availableTimeSlots prop
The AppointmentButton component has been updated to use the availableTimeSlots prop instead of the doctor_id prop. This change allows the component to receive an array of time slots and display them in the UI. The selectDoctor state variable has also been updated to store an array of TimeSlot objects. This enhancement improves the functionality of the component and ensures that the correct time slots are shown for booking appointments.